CF 1.6
-New: Histogram Adjust.
-Fixed: Kuwahara.
-Fixed: Fake HDR.

+++ Catalano.Statistics/src/Catalano/Statistics/Histogram.java
@@ -75,7 +75,7 @@ public class Histogram {
         P[0] = (double) c / n;
         for (int i = 1; i < length; i++) {
             c += values[i];
-            P[i] = (double) c / n;
+            P[i] = (double) c / (double)n;
         }
         
         return P;
